+++ R/ess_all_cntrounds.R
@@ -8,9 +8,10 @@
 #' 
 #' @param country A character of length 1 with the full name of the country.
 #'  Use \code{\link{show_countries}} for a list of available countries.
-#' @param your_email a character vector with your email, such as "your_email@email.com".
+#' @param ess_email a character vector with your email, such as "your_email@email.com".
 #' If you haven't registered in the ESS website, create an account at 
-#' \url{http://www.europeansocialsurvey.org/user/new}
+#' \url{http://www.europeansocialsurvey.org/user/new}. A prefered method is to login
+#' through \code{\link{set_email}}.
 #' @param only_download whether to only download the files as Stata files. Defaults to FALSE.
 #' @param output_dir a character vector with the output directory in case you want to only
 #' download the files using the \code{only_download} argument. Defaults to your working directory.
@@ -32,8 +33,10 @@
 #' @examples
 #' \dontrun{
 #' 
+#' set_email("your@email.com")
+#' 
 #' # Will return all available rounds for Denmark
-#' ess_all_cntrounds("Denmark", "your_email@gmail.com")
+#' ess_all_cntrounds("Denmark")
 #' 
 #' # Will download all rounds to the directory stored below
 #' # as stata files (set by default) and won't return the rounds
@@ -43,19 +46,18 @@
 #' 
 #' ess_all_cntrounds(
 #' "Denmark",
-#' "your_email@gmail.com",
 #'  only_download = TRUE,
 #'  output_dir = dl_dir
 #'  )
 #' 
 #' }
-ess_all_cntrounds <- function(country, your_email, only_download = FALSE,
+ess_all_cntrounds <- function(country, ess_email = NULL, only_download = FALSE,
                               output_dir = getwd(), format = 'stata') {
 
     ess_country(
     country,
     show_country_rounds(country),
-    your_email,
+    ess_email,
     only_download,
     output_dir,
     format